chore: move optimism types to their own file
